# docs/swagger.yaml
basePath: /
definitions:
  errors.DefaultError:
    properties:
      message:
        type: string
    type: object
  errors.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  models.Client:
    properties:
      active:
        type: boolean
      email:
        maxLength: 125
        type: string
      id:
        type: string
      name:
        maxLength: 127
        minLength: 3
        type: string
      surname:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - email
    - name
    - surname
    type: object
host: localhost:8000
info:
  contact: {}
  description: Client registration service that enable to manage clients, their addresses,
    their contacts and their documents. In this service is also included the client
    registration
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://posteris.io/terms/
  title: Client Registration service
  version: "1.0"
paths:
  /api/v1/clientes:
    get:
      consumes:
      - application/json
      description: List all clients
      parameters:
      - description: Client Name
        in: query
        name: name
        type: string
      - description: Client Name
        in: query
        name: surname
        type: string
      - description: Client email
        format: email
        in: query
        name: email
        type: string
      - description: is cliente active?
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.DefaultError'
      summary: List Clients
      tags:
      - Client
    post:
      consumes:
      - application/json
      description: Create a new client and persist it at the database.
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Client'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            items:
              $ref: '#/definitions/errors.ValidationError'
            type: array
        "415":
          description: Unsupported Media Type
          schema:
            $ref: '#/definitions/errors.DefaultError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.DefaultError'
      summary: Create new Client
      tags:
      - Client
  /api/v1/clientes/{id}:
    get:
      consumes:
      - application/json
      description: Obtains the client object based in their ID
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.DefaultError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.DefaultError'
      summary: Get client by ID
      tags:
      - Client
swagger: "2.0"